Split dump_images.go
